refactor: move VoteState to session package and introduce VoteStateDTO for frontend
